requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.24
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
